extract statistics
