+++ src/App/index.tsx
@@ -4,9 +4,9 @@ import { createEffect, createMemo, createSignal, For, Match, onCleanup, onMount,
 import { createStore, produce, SetStoreFunction } from 'solid-js/store';
 import { Portal } from 'solid-js/web';
 import superjson from 'superjson';
-import pomodoroSvg from './pomodoro.svg';
 import { getDefaultDataStore, Item, DataStore, DataContext } from './store/data';
 import { AppContext, getDefaultAppStore } from './store/app';
+import { Statistics } from './Statistics';
 
 // component
 export function App() {
@@ -38,86 +38,6 @@ export function App() {
     onCleanup(() => clearInterval(intervalId));
   });
 
-  // stats
-  const [statTime, setStatTime] = createSignal<'day' | 'week' | 'month' | 'year' | 'all'>('day');
-  const statTimeStartDate = createMemo(() => {
-    const from = selectedDate();
-    const time = statTime();
-
-    switch (time) {
-      case 'day':   return from;
-      case 'week':  return getStartOfWeek(from);
-      case 'month': return new Date(from.getFullYear(), from.getMonth(), 1);
-      case 'year':  return new Date(from.getFullYear(), 0, 1);
-      case 'all':   return new Date(0);
-      default:      throw new Error(time satisfies never);
-    }
-  });
-
-  function dateFilter(item: Item) {
-    const target = statTimeStartDate();
-    const itemDate = getDateNoTime(item.start);
-    const time = statTime();
-
-    switch (time) {
-      case 'day':
-        return itemDate.toDateString() === target.toDateString();
-      case 'week':
-        return getStartOfWeek(itemDate).toDateString() === target.toDateString();
-      case 'month':
-        return itemDate.getFullYear() === target.getFullYear()
-            && itemDate.getMonth() === target.getMonth();
-      case 'year':
-        return itemDate.getFullYear() === target.getFullYear();
-      case 'all':
-        return true;
-      default:
-        throw new Error(time satisfies never);
-    }
-  }
-
-  const [sortBy, setSortBy] = createSignal<'tag' | 'duration' | 'pomodoros'>('tag');
-  const [sortOrder, setSortOrder] = createSignal<'asc' | 'desc'>('asc');
-
-  function changeSorting(by: 'tag' | 'duration' | 'pomodoros') {
-    if (sortBy() === by) {
-      setSortOrder(sortOrder() === 'asc' ? 'desc' : 'asc');
-    } else {
-      setSortBy(by);
-      setSortOrder('asc');
-    }
-  }
-
-  const dayStats = createMemo(() => calculateStatsAtDate(
-    dataStore.items,
-    dateFilter,
-  ));
-
-  const sortedStats = createMemo(() => {
-    const { entries, sumAll } = dayStats();
-
-    const stats = entries.map(item => ({
-      ...item,
-      pomodoros: toPomodoro(item.duration),
-    }));
-
-    stats.sort((a, b) => {
-      const aVal = a[sortBy()];
-      const bVal = b[sortBy()];
-
-      if (typeof aVal === 'string' && typeof bVal === 'string') {
-        return sortOrder() === 'asc'
-          ? aVal.localeCompare(bVal)
-          : bVal.localeCompare(aVal);
-      }
-
-      return sortOrder() === 'asc'
-        ? Number(aVal) - Number(bVal)
-        : Number(bVal) - Number(aVal);
-    });
-
-    return { entries: stats, sumAll };
-  });
 
   const allTags = createMemo(() => {
     const tags = dataStore.items.map(item => item.tag);
@@ -288,17 +208,6 @@ export function App() {
     if (data) setDataStore(data);
   }
 
-  function getWeekInterval(date: Date) {
-    const start = getStartOfWeek(date);
-    const end = new Date(start);
-    end.setDate(start.getDate() + 6);
-
-    const startStr = start.toLocaleDateString('en-US', { month: '2-digit', day: '2-digit' });
-    const endStr = end.toLocaleDateString('en-US', { month: '2-digit', day: '2-digit' });
-
-    return `${startStr} - ${endStr}`;
-  }
-
   return (
     <AppContext.Provider value={[appStore, setAppStore]}>
       <DataContext.Provider value={[dataStore, setDataStore]}>
@@ -409,84 +318,7 @@ export function App() {
 
           <br />
           Statistics
-          <div class={sToolbar}>
-            <div class={sToolbarLeft}>
-              <label>
-                <input type="radio" name="timeRange" value="day"
-                  onChange={() => setStatTime('day')}
-                  checked={statTime() === 'day'}
-                />
-                Day ({selectedDate().toLocaleDateString('en-US', { month: '2-digit', day: '2-digit' })})
-              </label>
-              <label>
-                <input type="radio" name="timeRange" value="week"
-                  onChange={() => setStatTime('week')}
-                  checked={statTime() === 'week'}
-                />
-                Week ({getWeekInterval(selectedDate())})
-              </label>
-              <label>
-                <input type="radio" name="timeRange" value="month"
-                  onChange={() => setStatTime('month')}
-                  checked={statTime() === 'month'}
-                />
-                Month ({selectedDate().toLocaleDateString('en-US', { month: 'long' })})
-              </label>
-              <label>
-                <input type="radio" name="timeRange" value="year"
-                  onChange={() => setStatTime('year')}
-                  checked={statTime() === 'year'}
-                />
-                Year ({selectedDate().toLocaleDateString('en-US', { year: 'numeric' })})
-              </label>
-              <label>
-                <input type="radio" name="timeRange" value="all"
-                  onChange={() => setStatTime('all')}
-                  checked={statTime() === 'all'}
-                />
-                All time
-              </label>
-            </div>
-          </div>
-
-          <div class={sTableStats}>
-            <div class={sRow}>
-              <div class={cx(sCell, sCellHeader)} onClick={() => changeSorting('tag')}>Tag</div>
-              <div class={cx(sCell, sCellHeader)} onClick={() => changeSorting('duration')}>Duration</div>
-              <div class={cx(sCell, sCellHeader)} onClick={() => changeSorting('pomodoros')}>Pomodoros (30 min)</div>
-            </div>
-            <For each={sortedStats().entries}>
-              {(entry) => (
-                <div class={sRow}>
-                  <div class={sCell}>{entry.tag}</div>
-                  <div class={sCell}>{minutesToHoursMinutes(entry.duration)}</div>
-                  <div class={cx(sCell, sCellPomodoro)}>
-                    <Show when={entry.pomodoros > 0}>
-                      <Switch>
-                        <Match when={entry.tag === 'idle'}>
-                          <span>🌞 🌴 ⛱️ 🧘‍♀️ 🍹</span>
-                        </Match>
-                        <Match when={Math.floor(entry.pomodoros) > 4}>
-                          <PomodoroIcon /> x{Math.floor(entry.pomodoros)}
-                        </Match>
-                        <Match when={Math.floor(entry.pomodoros) <= 4}>
-                          <For each={Array(Math.floor(entry.pomodoros))}>
-                            {() => <PomodoroIcon />}
-                          </For>
-                          <PomodoroIcon amount={entry.pomodoros % 1} grayed={true} />
-                        </Match>
-                      </Switch>
-                    </Show>
-                  </div>
-                </div>
-              )}
-            </For>
-            <div class={sRow}>
-              <div class={cx(sCell)}></div>
-              <div class={cx(sCell)}><b>{minutesToHoursMinutes(sortedStats().sumAll)}</b></div>
-              <div class={cx(sCell)}></div>
-            </div>
-          </div>
+          <Statistics />
 
           <br />
           Utilities
@@ -517,19 +349,6 @@ function updateTimestamp(date: Date, timestamp: string) {
   }
 
   return newDate;
-}
-
-function PomodoroIcon(props: { amount?: number, grayed?: boolean }) {
-  const targetWidth = () => 24 * (props.amount ?? 1);
-  const width = () => targetWidth() >= 10 ? targetWidth() : 0;
-
-  return <img
-    width={width().toFixed(2)}
-    height={24}
-    src={pomodoroSvg}
-    alt="Pomodoro"
-    classList={{ [sPomodoroGrayed]: props.grayed }}
-  />;
 }
 
 // api
@@ -623,56 +442,6 @@ function calculateDuration(start: Date, end: Date) {
   return Math.ceil((end.getTime() - start.getTime()) / (1000 * 60));
 }
 
-function minutesToHoursMinutes(minutesAmount: number) {
-  const hours = Math.floor(minutesAmount / 60);
-  const minutes = minutesAmount % 60;
-
-  if (hours > 0) {
-    return `${hours} h ${minutes} min`;
-  }
-
-  return `${minutes} min`;
-}
-
-function calculateStatsAtDate(itemsAll: Item[], filter: (item: Item) => boolean) {
-  const itemsAtDate = itemsAll.filter(filter);
-
-  const tags = [...new Set(itemsAtDate.map(item => item.tag))];
-
-  const entries = tags.map(tag => {
-    const now = new Date();
-    const items = itemsAtDate.filter(item => item.tag === tag);
-    const duration = items
-      .reduce((sum, item) => sum + calculateDuration(item.start, item.end ?? now), 0);
-
-    return { tag: tag || '*empty*', duration };
-  });
-
-  const sumAll = entries.reduce((sum, entry) => sum + entry.duration, 0);
-
-  return { entries, sumAll };
-}
-
-function getDateNoTime(date: Date) {
-  const target = new Date(date);
-  target.setHours(0, 0, 0, 0);
-
-  return target;
-}
-
-function toPomodoro(minutes: number) {
-  return minutes / 30;
-}
-
-function getStartOfWeek(date: Date) {
-  const result = new Date(date);
-  const day = result.getDay();
-  const diff = day === 0 ? -6 : 1 - day; // Adjust to make Monday the first day
-  result.setDate(result.getDate() + diff);
-  result.setHours(0, 0, 0, 0);
-  return result;
-}
-
 function triggerNonDestructiveBlur(e: KeyboardEvent & { currentTarget: HTMLDivElement }) {
   const selection = window.getSelection();
   const offset = selection?.focusOffset || 0;
@@ -698,6 +467,7 @@ const sApp = css`
   min-width: 500px;
 `;
 
+// toolbar
 const sToolbar = css`
   display: flex;
   justify-content: space-between;
@@ -717,6 +487,7 @@ const sToolbarRight = css`
   align-items: center;
 `;
 
+// ...
 const sCurrentDate = css`
   display: flex;
   justify-content: space-between;
@@ -787,15 +558,6 @@ const sCellEditableText = css`
 const sCellGrayed = css`
   color: #555;
   font-style: italic;
-`;
-
-const sCellPomodoro = css`
-  display: flex;
-  gap: 5px;
-`;
-
-const sPomodoroGrayed = css`
-  filter: grayscale(100%) brightness(120%);
 `;
 
 const sTagList = css`
